requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pyTelegramBotAPI==4.14.0
google-genai>=0.3.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
pydantic-settings==2.1.0
pyjwt==2.8.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
schedule==1.2.1
Pillow==10.0.0
psycopg2-binary==2.9.9
fpdf2==2.8.5
matplotlib==3.8.2
pandas==2.1.3
slowapi==0.1.7


